backend/colorizer/ml_model.py
import numpy as np
import cv2
from skimage import color
import os
import tensorflow as tf
from tensorflow.keras.models import Sequential, Model
from tensorflow.keras.layers import Conv2D, UpSampling2D, InputLayer, Conv2DTranspose
from tensorflow.keras.preprocessing.image import img_to_array, load_img

class ColorizationModel:
    """
    Class for handling image colorization using a deep learning model.
    This is a simplified placeholder class for development.
    """
    def __init__(self):
        self.model = None
        self.input_shape = (256, 256, 1)  # Default input shape
        self.initialized = False
    
    def build_model(self):
        """
        Placeholder for model building
        """
        self.initialized = True
        return self.model
    
    def colorize(self, image_path, output_path):
        """
        Placeholder for colorization process
        """
        # In a real implementation, this would colorize the image
        # For now, this is just a placeholder
        
        # Create a dummy file or just return the path
        with open(output_path, 'w') as f:
            f.write('Placeholder for colorized image')
        
        return output_path
    
    def train(self, dataset_path, epochs=10, batch_size=32):
        """
        Placeholder for model training
        """
        print(f"[Mock] Training model on {dataset_path} for {epochs} epochs with batch size {batch_size}")
        return True 